Make winit an opt-in feature of egui-wgpu
